Add a multi-line subtitle widget
This adds a multi-line subtitle widget based on the SubtitleTextblock widget.
The widget can be styled in the same way as the SubtitleTextblock, but this
provides the option to show several lines of subtitles at once, and it handles
timing out subtitles after they stop being emitted.

Please see the note in the header regarding potential i18n issues with this
widget.