Add a vertical box that permits several subtitles to appear simultaneously
This is intended to aid in showing multiple subtitles at the same time
to improve accessibility for players who have hearing issues: with the
default subtitle display, only one subtitile is shown at once, even if
more than one potentially meaningful sound is playing. This allows
several lines of subtitle to appear simultaneously so that potentially
important sound indicators are not lost.